renamed PageDelegates to PageStaking
