+++ app/src/renderer/components/staking/PageStaking.vue
